+++ tests/use_cases/beam_model/test_constraints.py
@@ -56,9 +56,8 @@ def test_default_outputs(discipline):
     """Check the default values of the outputs."""
     assert_almost_equal(
         discipline.local_data["c_stress"],
-        discipline.local_data["sigma_all"] / (discipline.local_data["sigma_vm"] + 1.0),
+        discipline.local_data["sigma_vm"] / discipline.local_data["sigma_all"],
     )
     assert_almost_equal(
-        discipline.local_data["c_displ"],
-        100.0 / (discipline.local_data["displ"] + 0.1),
+        discipline.local_data["c_displ"], discipline.local_data["displ"] / 100.0
     )